requirements.txt
loguru==0.7.0
PyGithub==1.58.1